==> scripts/04_VerifyFAIMSrs.Rmd <==
---
title: "Merging 2017-2018 data with RS"
author: "Adela Sobotkova"
date: "24 June 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# GOAL, INPUTS AND OUTPUTS

As input this script takes
1) 2017 and 2018 dataset from FAIMS, streamlined in OR, created with 01_LoadData.R,
2) mounds_RS which is the product of remote sensing in GoogleEarth Pro of map features digitized in 2010. It contains landuse categories, which might help with topo_mounds


The main aim here is to merge the two inputs and look at the consistency of landuse between the two datasets. Difference helps map the gap between subjective perspective of a mound visitor, who describes the landuse as s/he sees it in the immediate environment versus the objective? view of a RS specialist who looks at the landuse from bird's eye perspective and can review it over time at different scales.
Furthermore, difference may also collate some errors of English ("perennial" for example was a term whose meaning differed among the teams), errors of agricultural knowledge (fallow fields were often interpreted as 'pastures'), and errors of observation (swapping landuse on top with landuse around, relevant only in 2009-2010, though). 
Conceptual differences also intercede: mounds rendered through agriculture to be low and extensive features with a small lozenge of grass covered stones on the top often have fuzzy boundaries. In 2010 such mounds were described as smaller features (15m in diameter) delimited by the grassy patch (LU_top = grass), while in 2017 the entire extent of the mound (e.g. 50x50m base with a grassy patch on top) was taken into account and dominant LU on top would have been annual agriculture. 
The FAIMS data on landuse 2017-2018 was collected with the help of picture dictionaries so should be devoid of errors of English and of observation, but may still suffer from insufficient agricultural knowledge and contains the shift in conceptual understanding of a 'mound'. 

# MERGE by Trap (Is there TopoID in 2017-2018 mnds?)
```{r}
topo_rs <- left_join(mnd2017, mounds_RS, 
                         by = c("identifier"=""), copy = FALSE, suffix = c(".topo", ".rs")) # 
```

CONTINUE...

```{r}
topors <- topo_rs %>% 
  select(TopoID, Trap, TRAPCode, SomethingPresentOntheGround, Type, LandUseAround, LanduseOn, LU_Around, LU_top, Length.topo, `Length (max, m)`, Length.rs, RS_Status, RS_feature_type, Condition, Principal)

```

How many of the original 850 RS features have a counterpart in the Verified Mound data?

```{r}
topors %>%
  filter
  filter(!is.na(RS_Status))
```
