==> scripts/06b_SpatialEnrichment.R <==
##################################################################
## Mound data enrichment from available raster and spatial data
# created 29 July 2020
# Adela Sobotkova
##################################################################

# Goal
# Enrich mound dataset with elevation, slope and aspect, and calculate distance to nearest settlement boundary as well as to the national border.
# As inputs we use the 
# - shapefile dataset from 06_SpatialJoin.R, 
# - ASTER DEM from JICA 1994, and 
# - vector files of settlement polygons for Bulgaria, Yambol with 15km buffer, and national border provided by JICA
# - table of distances from master_sp to nearest settlement generated in ArcGIS 
# 
# Process
# 1.Load data
# 2.Check CRS everywhere for alignment with EPSG 32635
# 3.Sample rasters at points: I extract environmental data from rasters at mound coordinates by using
# the raster::extract() function
# 4. Calculate distance indeces via the  st_nearest_feature() and st_distance()
# 5. Calculate mound prominence/hierachy via raster::extract() with perc() function
# 6. OUPUT: interim dataset enviro_Yamnds, a dataframe with environmental indeces of 
# elevation, slope, aspect, distance to national border, distance to nearest settlement, prominence at different radii.

# Libraries
library(tidyverse)
library(raster)
library(rgdal)
library(FSA)
library(sf)

################################# LOAD INPUTS   ###########################################################################

# Load ASTER raster, in this case ASTER DEM from JICA, projected to WGS82, UTM 35N

#Yam <- raster("E:/TRAP Workstation/Shared GIS/Satellite imagery/ASTER/DEM/ASTGTM_N42E026/prjYAM_DEM_N42E026.tif")
Yam <- raster("F:/Shared GIS/Satellite imagery/ASTER/DEM/ASTGTM_N42E026/prjYAM_DEM_N42E026.tif")
Yam # projected WGS84 35N raster 27m resolution, 4112x3053
#Yaspect<- raster("E:/TRAP Workstation/Shared GIS/Satellite imagery/ASTER/DEM/ASTGTM_N42E026/prjYAM_Aspect.tif")
#Yslope<- raster("E:/TRAP Workstation/Shared GIS/Satellite imagery/ASTER/DEM/ASTGTM_N42E026/prjYAM_Slope.tif")


# Load shapefiles of master_sp, urban areas, boundaries etc. generated by ArcGIS (1 spatial duplicate)

#mnd_shp <- st_read("C:/Users/Adela/Documents/Professional/Projects/MQNS/GIS/Vectors/200918VisitedMax.shp")
#Y_region <- st_read("F:/Shared GIS/Vector/Oxbow Shapefiles/13_Janouch/Area_borders_shp/Yambol_province_TRAP.shp")

bg_border <- st_read("F:/Shared GIS/Vector/Oxbow Shapefiles/Bulgaria_Border.shp")
Y_towns <- st_read("F:/Shared GIS/Vector/Oxbow Shapefiles/13_Janouch/Rivers_Roads_Settlements_Yambol/Modern_settlements_Yambol_TRAP.shp")
towns <- st_read("C:/Users/Adela/Documents/Professional/Projects/MQNS/GIS/Vectors/Yamurbanarea.shp") # towns within 15km buffer of Yambol boundary exported from JICA


# Load distance to settlement ("near") table from ArcGIS if not wanting to deal with sf package

distanceTown <- read_csv("C:/Users/Adela/Documents/Professional/Projects/MQNS/GIS/Vectors/MndTownDistance.csv")
distanceTown <- distanceTown %>%
  select(TRAP,NEAR_DIST, NEAR_X, NEAR_Y)

################################# VISUAL CHECK - PLOT LOADED DATASETS ##########################################################################

# Visual check of the data
plot(Yam)
plot(master_sp$geometry, pch = 17, add = TRUE)
plot(towns$geometry, add = TRUE)


plot(bg_border$geometry)
plot(towns$geometry, add = TRUE)
plot(master_sp$geometry, pch = 17, add = TRUE)

################################# CHECK CRS AND GRAB MOUND COORDINATES ##########################################################################

# Prerequisites to sampling - Coordinate system alignment

crs(Yam)
crs(master_sp)
crs(Y_towns)
crs(bg_border)
st_crs(Y_region)
all.equal(st_crs(Yam),st_crs(master_sp)) # LOOKS GREAT
all.equal(st_crs(bg_border),st_crs(master_sp)) # LOOKS GREAT
all.equal(st_crs(Y_towns),st_crs(master_sp)) # LOOKS GREAT
all.equal(st_crs(towns),st_crs(master_sp))

st_coordinates(master_sp)
mound_coordinates <- data.frame(st_coordinates(master_sp)) # don't add TRAP yet

################################# CALCULATE DISTANCE TO BG BORDER AND NEAREST TOWN ###########################################################################

# Distance from points to BG boundary

enviro_Yammnds$distBG <- st_distance(master_sp,bg_border)
enviro_Yammnds[1:3,]


# Distance from points to nearest settlement

?st_nearest_feature()

# R - way - BOTTOM EXAMPLE WORKS! 
# enviro_Yammnds$distTown <- st_distance(master_sp,towns) # creates a list which is unusable at moment

# Following Hadley's comment from 2010 on 
# https://stackoverflow.com/questions/4512465/what-is-the-most-efficient-way-to-cast-a-list-as-a-data-frame
# I should use 'plyr::ldply(my.list, data.frame)', but it seems slow and superceded by do.call
# test2 <- do.call(rbind, as_tibble(enviro_Yammnds$distTown)) # takes forever

#https://stackoverflow.com/questions/49200458/find-nearest-features-using-sf-in-r (does not work for me)


# Get distance from mound to nearest town by 
# (1) selecting closest town via st_nearest and (2) getting distance with st_distance
nearest <-  try(st_nearest_feature(master_sp, towns))
distSettle <- st_distance(master_sp, towns[nearest,], by_element = TRUE)

dist[2:5]
distanceTown[2:5,2]
distanceTown

# add result to dataframe
class(enviro_Yammnds)
dim(enviro_Yammnds)
enviro_Yammnds$distTown <- st_distance(master_sp, towns[nearest,], by_element = TRUE)
head(enviro_Yammnds,3)

################################# SAMPLE ELEVATIONS, ASPECT AND SLOPE ###########################################################################

# Sample elevations at mound locations

elevation <- raster::extract(Yam, mound_coordinates)
elev_Yammnds <- data.frame(mound_coordinates, master_sp$TRAP,elevation=
                             raster::extract(Yam, mound_coordinates))

# Create a table with elevation, slope and aspect at mound locations

slope <- raster::extract(Yslope, mound_coordinates)
enviro_Yammnds <- data.frame(mound_coordinates, master_sp$TRAP,
                             elevation=
                               raster::extract(Yam, mound_coordinates),
                             slope=
                              raster::extract(Yslope, mound_coordinates),
                             aspect=
                               raster::extract(Yaspect, mound_coordinates))



################################# PROMINENCE ###################################

# There is something rotten in the prominence below, as it always averages out to a random 50% (unlikely) ; when I tested with a subset 
# the numbers were closer to my epxectations, so maybe it's a lot of master_sp on a slope. 
# PROBLEM IS CONCEPTUAL: PROMINENCE IS ABOUT SPATIAL HIERARCHY, WHILE i WAS THINKING MORE OF VISIBILITY.
enviro_Yammnds$prom250mbuff <- raster::extract(Yam,  # raster containing elevation data
                                        mound_coordinates, # centroids of master_sp
                                        buffer = 250,
                                        #buffer = 250, # actual buffer size in crs units, in this case 250m  or ca 22x22 cells around kernel
                                        fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})



enviro_Yammnds%>% 
  filter(prom250mbuff > 74) %>% 
  plot()


# Test with smaller subset to verify that the prominence values vary and are not random
mnd09shp <- master_sp %>% 
  filter(TRAP%in%mnd2009$TRAP)%>% 
  filter(TRAP < 8020)

mnd09_coordinates <- data.frame(st_coordinates(mnd09shp)) # don't add TRAP yet
test <- raster::extract(Yam,  # raster containing elevation data
                mnd09_coordinates, # centroids of master_sp
                buffer = 250,
                #buffer = 250, # actual buffer size in crs units, in this case 250m  or ca 22x22 cells around kernel
                fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})
summary(test)
hist(test)


# Prominence at 1 and 2 and 5 km

enviro_Yammnds$prom1kmbuff <- raster::extract(Yam,  # raster containing elevation data
                                               mound_coordinates, # centroids of master_sp
                                               buffer = 1000, # actual buffer size in crs units, in this case 1000m  or ca 50x50 cells around kernel
                                               fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})

enviro_Yammnds$prom2kmbuff <- raster::extract(Yam,  # raster containing elevation data
                                               mound_coordinates, # centroids of master_sp
                                               buffer = 2000,# actual buffer size in crs units, in this case 2000m  or ca 100x100 cells around kernel
                                               fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})

enviro_Yammnds$prom3kmbuff <- raster::extract(Yam,  # raster containing elevation data
                                              mound_coordinates, # centroids of master_sp
                                              buffer = 3000,# actual buffer size in crs units, in this case 3000m  or ca 200x200 cells around kernel
                                              fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})

enviro_Yammnds$prom5kmbuff <- raster::extract(Yam,  # raster containing elevation data
                                              mound_coordinates, # centroids of master_sp
                                              buffer = 5000,# actual buffer size in crs units, in this case 5000m  or ca 400x400 cells around kernel
                                              fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})

enviro_Yammnds$prom10kmbuff <- raster::extract(Yam,  # raster containing elevation data
                                              mound_coordinates, # centroids of master_sp
                                              buffer = 10000,# actual buffer size in crs units, in this case 10000m  or ca 500X500 cells around kernel
                                              fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})


hist(enviro_Yammnds$prom2kmbuff)


################################# CREATE AGGREGATE DATASET ################################################################

# join DISTANCE TO TOWN  to the master dataset
masterall <- master %>% left_join(distanceTown, by = "TRAP")
masterall # dataset suitable for glm(), as it has effect, landuse, distance to town, etc.
masterall$effect


# join DISTANCE TO BORDER and other ENVIRONMENAL data to the master dataset
masterall <- masterall %>% left_join(enviro_Yammnds[,1:10], by = c("TRAP" = "master_sp.TRAP"))

################################# VISUALIZE SOME RESULTS ########################################################################## 
# PREPARE the RASTER

# Cropping a raster using a vector
# We’ll use the rgdal package to read the clip shapefile. 
# The first argument is the folder location (relative to the session’s working directory) 
# and the second argument is the name of the shapefile (without the .shp extension).
library(rgdal)
shp   <- readOGR(".","Clip")
elev2 <- crop(elev.mem, shp)

# Cropping a raster interactively
# You can also crop a raster manually by drawing either a rectangle or a polygon directly on the plot.
plot(Yam)
Yam2 <- select(Yam, use="rec") # To select by rectangle: select topleft and bottomright corner!
plot(Yam2)
plot(shapefile$geometry, add = TRUE)
Yam2 <- select(Yam, use="pol") # To select by polygon, press escape to finish drawing the polygon


# VIEW RASTER VALUES
Yam_df <- as.data.frame(Yam, xy = TRUE)


# Clip mound data
plot(shapefile)
plot(st_intersection(shapefile$geometry, Y_region))
plot(Y_region, add = TRUE)

nrow(elev_Yammnds)
hist(Yam_df$prjYAM_DEM_N42E026, main="Overlapping Regional and Mound Elevation")
hist(elev_Yammnds$elevation/rows, col=rgb(0.8,0.8,0.8,0.5))#, add=TRUE)
box()

ggplot() +
  geom_histogram(data=Yam_df, aes(prjYAM_DEM_N42E026))



Yam_df <- Yam_df %>%
  mutate(fct_elevation = cut(prjYAM_DEM_N42E026, breaks = 10))

ggplot() +
  geom_bar(data = Yam_df, aes(fct_elevation)) +
  geom_bar(elev_Yammnds, aes(elevation)) # fix this second line

plot(Yam)
plot(master_sp$geometry, col = master_sp$prom1000mbuff, add = TRUE)

# Sample prominence at mound locations
# https://geocompr.robinlovelace.net/spatial-operations.html#spatial-raster-subsetting
# https://mgimond.github.io/megug2017/
