==> scripts/07_Finalize.Rmd <==
---
title: "Final cleaning and linking of AKB IDs to Yambol Features"
author: "Adela Sobotkova"
date: "updated `r Sys.Date()`"
output:
  rmdformats::readthedown:
    highlight: kate
---

```{r setup, include=FALSE}
library(knitr)
library(rmdformats)
library(formatR)

## Global options
options(max.print="75")
opts_chunk$set(echo=TRUE,
	             cache=TRUE,
               prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

The script explores the consistency of datatypes and feature attributes created in previous scripts, streamlines loose ends, and marries the records to their AKB counterparts.
Final data are clipped to Yambol boundary and spatial duplicates are identified and filtering facilitated by adding logical campaign-indicating columns.
This script can be run standalone and is final. There is no need to run previous scripts unless one wishes to explore the data in their raw state or change some of the processing.

## What is AKB 

AKB stands for the "Arheologicheska Karta na Bulgaria", national digital register, where all archaeological sites and places of interest are entered upon their first investigation. The purpose of the present script is to add AKB identifiers assembled by Todor Valchev to the Yambol field observations. The biggest benefit of this exercise is that the AKB reference lets us see if a given mound was excavated and dated. It also allows users with access to AKB to look up the full cultural heritage information associated with the record.

Sites are often entered into AKB upon their first visit, with additional data added upon revisit, excavations, or other analyses.

The AKB link is one reason why you may want to filter the Yambol data for one or another spatial versions: `Upto2010` and `Post2010`. While the `Post2010` is based on later revisits and thus more recent and authoritative (especially if driven by the need to update or improve the record), the `Upto2010` record IDs were use to create links in AKB and so are more suitable for mapping of cultural heritage data in AKB with the environmental variables present here.

## Setup
```{r libraries}
library(tidyverse)
library(sf)
library(mapview)
```


## Load data: `master_sp_enriched`
Here we load aggregated and enriched features from Yambol and wider surroundings. 
```{r load-aggr-data}
# not deduplicated for final dataset
features <- st_read("../output_data/interim/master_sp_enriched.geojson")
```

## Validation I: Check how many of the features are actually mounds

```{r type-tally}
# Filter features by type
features %>% 
  group_by(Type) %>% 
  tally()
```

```{r type-check}
# Verify the Source in the category "Other"  of Type.

features %>% 
  filter(Type == "Other") %>% 
  group_by(Source) %>% 
  tally()

```
In the tally of "Other" types, there are 34 Legacy verification features and 41 Survey features in the later monitoring and 30 LgV and 41 Survey in the earlier monitoring. While the latter are expected, the 34 verified features required follow up on 27 Dec 2022. Inspection showed that many of the verified features originate not from sunbursts but other map markers, such as rayed squares and triangles. We inspected these in the early seasons but stopped doing so after these turned out not to lead to mounds reliably. 

### What is hiding under "other"?
```{r explore-type-other}
features %>% 
  filter(Type == "Other" & Source == "Legacy verification")  %>% 
  group_by(PrincipalSourceOfImpact) %>% 
  tally()
```
We can see that a lot of the map symbols led to geodetic points and benchmarks, 8 were not accessible, and other locations contained water wells and other structures, whose setting prevented us from labelling them as extinct mounds

## Validation II: Check for attribute duplicates 
Spatial duplication is addressed in 06_Enrich.Rmd
```{r duplicate-check}
features$TRAP[duplicated(features$TRAP)] 
```

## Validation III: Clean up condition, height and other attribute

Condition was documented on Likert scale from 1 to 5 with 1 being pristine and 5 being extinct. 0 denoted No `observation`.
```{r clean-up-condition}
unique(features$Condition)
glimpse(features)

features <- features %>%
   dplyr::mutate(Condition = case_when(Condition == 0 ~ "NA",
                               Condition == 6 ~ "5",
                               Condition != 0 ~ Condition)) %>% 
  dplyr::mutate(Condition = as.factor(Condition)) %>% 
  dplyr::mutate(TypeCertainty = case_when(
    grepl("\\?",Type) ~ "Uncertain",
    !grepl("\\?",Type) ~ "Certain")) %>% 
  dplyr::mutate(Type= gsub("\\?","",Type)) %>% 
  dplyr::mutate(DiameterMax = as.numeric(DiameterMax)) 

levels(features$Condition) = c(1,2,3,4,5, NA)

features %>% 
  group_by(TypeCertainty) %>% 
  tally()
```

## Fix I: Constrain to Yambol Province 

```{r yambol-features}
# load the boundary
Y_region <- st_read("../input_data/Vectors/YamRegion.shp") 

Y_features <- st_intersection(features, Y_region$geometry)
Y_features  # 1260 features and 24 fields
```

Within Yambol, we have 1260 observations at map locations. 


#### Add zemlishte information

```{r add-zemlishte-to-Yambol}
zem <- st_read("../input_data/Vectors/Yambol_zemlishta.geojson")
mapview(zem$geometry)
Y_features <- st_join(Y_features, zem[,"Name_en"], join = st_intersects)
```


## Fix II: Implement changes following manual review
This chunk corrects poor data entries in the structured digital forms on the basis of photographs, diaries, and AKB records.

```{r changes-type-source}
# Type Certainty UNCERTAIN TO CERTAIN
certain <- Y_features %>% 
  filter(TRAP %in% c(9838,9852,8766,8359)) %>% 
  mutate(TypeCertainty = "Certain" )

# Type Burial mound to EXTINCT
extinct <- Y_features %>% 
  filter(TRAP %in% c(9516, 8337, 8388, 8770, 9910, 9911)) %>% 
  mutate(Type = "Extinct Burial Mound" )

# Type Burial mound to Scatter
scatter <- Y_features %>% 
  filter(TRAP == 9883) %>% 
  mutate(Type = "Surface Scatter" )

# Type Burial mound to Other
other <- Y_features %>% 
  filter(TRAP %in% c(8427, 9645))%>% 
  mutate(Type = "Other")

# Swap Source in these two and TypeCertainty
m8763 <- Y_features %>% 
  filter(TRAP == 8763) %>% 
  mutate(Source = "Legacy verification")
m8764 <- Y_features %>% 
  filter(TRAP ==  8764) %>% 
  mutate(Source = "Survey", Type = "Extinct Burial Mound", TypeCertainty = "Uncertain")
  
fixes <- rbind(m8763,m8764, other,scatter,extinct,certain)

class(fixes$distBG)

# Fix source in features with row_update()
colnames(Y_features)
colnames(fixes)

# row_update() fails on the units distBG and geometry columns so we eliminate these for a moment to apply fixes, and then rejoin them.  

features_fixed <- rows_update(
  #remove distBG
  st_drop_geometry(Y_features)[,-14],  
  st_drop_geometry(fixes)[,-14],
  by = "TRAP")

# rejoin geometry and distBG to cleaned source and type data
features_fixed <- features_fixed %>% 
  left_join(Y_features[,c("TRAP", "distBG")], by = "TRAP") %>% 
  st_as_sf()

features_fixed %>% mapview(zcol = "Type")
features_fixed %>% 
  dplyr::filter(Type == "Tell") %>% 
  dplyr::select(Date, TRAP, Source)
```
### Fix missing data notation
AllNotes field and TopoID field sometimes contain zeroes or are outright blank to indicate missing data. Missing data is legitimate, as only Map symbols get TopoID and many mounds were first encountered during Survey. Likewise, AllNotes field is aggregated from annotations in the mobile app, which served to comment on structured data (dropdown options), indicating some deviation or addition. They do not exist everywhere. 
Let's convert all these zeroes and blanks to NAs for consistency.
```{r missing-to-NAs}
features_fixed$AllNotes <- ifelse(features_fixed$AllNotes == "0" |
                                    features_fixed$AllNotes == "", NA, 
                                  features_fixed$AllNotes)
features_fixed$TopoID <- ifelse(features_fixed$TopoID == "0", NA, features_fixed$TopoID)
```


Clean up after the changes
```{r clean-up-objects}
rm(certain, extinct, fixes, m8763, m8764, other, scatter)
```

## Fix III: Add AKB identifiers
Here we add AKB numbers to feature data so their future analyses can be brought to bear on the environmental attributes documented during survey.

### 1055 AKB records
We have 1055 AKB numbers among 1466 visited features. The mismatch is because some features were not deemed worthy of registering (extinct or overbuilt status, military bunkers. etc)

```{r load-AKB}
AKB <- read_csv("../input_data/MoundsAKBnumbers.csv")
sum(!is.na(AKB$AKB))
names(AKB)
head(AKB)

# Which ones are excavated?
AKB %>% 
  filter(grepl("[Ee]xcavated", Notes)) 

# Two AKB numbers are assigned to one and the same TRAP id?
AKB %>%
  filter(TRAP == 9962)
which(AKB$TRAP == 9962)

# Eliminate duplicates
# AKB <- AKB %>%
#   slice(-968)
```


### Join AKB to features
```{r join-AKB}
Y_features <- features_fixed %>% 
  left_join(AKB, by = c('TRAP')) %>% 
  rename(AKBNotes = Notes)

mapview(Y_features, zcol= "AKBNotes")

colnames(Y_features)
head(Y_features[,10:27])

# Are there duplicate AKB numbers? 
AKBduplicated <- Y_features %>%
  filter(!is.na(AKB)) %>% 
  filter(duplicated(AKB)) %>% 
  pull(AKB)

Y_features %>% 
  dplyr::filter(AKB %in% AKBduplicated) %>% 
  dplyr::select(TRAP, AKB) %>% distinct(AKB)

Y_features <- Y_features %>% 
  # Group by identical columns
  group_by(across(-c(AKB, AKBNotes))) %>%  
  summarize(
    AKB = paste(unique(AKB), collapse = "; "),       # Combine AKBID
    AKBNotes = paste(unique(AKBNotes), collapse = "; ")  # Combine AKBNotes
    ) %>%
  ungroup() 
  # %>% 
  # filter(TRAP %in% 9962:9984) %>%
  # dplyr::select(starts_with("AKB"))

```

Beware: In 17 instances, AKB numbers are duplicated, meaning that some TRAP mounds share an AKB number. Most of these are spatial duplicates that received a new TRAP number upon revisit. We will deal with them below.

### How many features/mounds have been excavated in Yambol by 2023
```{r Yambol-exav-mnds}
Y_features %>% 
  filter(AKB > 0) %>% 
  filter(grepl("Excav", AKBNotes)) #  28 excavated in both datasets

Y_features %>% 
  filter(AKB > 0)  # 1030/1040 early/later features have AKB >> duplicates must be here
```

## Fix IV: Reorder columns and fix their names for legibility

```{r rename-reorder-colums}
names(Y_features)
Y_features <- Y_features %>% 
  dplyr::select(-Area) %>% 
  rename(TopoNote = Note, LanduseAround = LU_Around, LanduseOnTop = LU_Top, DiameterMax_m = DiameterMax, HeightMax_m = HeightMax, DistTown_m = distTown, DistTownBoundary_m = distTownBoundary, Prominence250m_percent = prom250mbuff, DistBG_m = distBG, AsterElev_m = ElevAster_m, AsterSlope_degrees = SlopeAster_degrees, AsterAspect = AspectAster, AsterTRI = TRI, AsterTPI = TPI, AsterRoughness = Roughness, Area = Name_en) %>% 
  relocate(c(TopoID, TopoNote), .after = PrincipalSourceOfImpact) %>% 
  relocate(Area, .after = TRAP) %>% 
  relocate(TypeCertainty, .after = Type)
```

## Fix V: Identify spatial duplicates

A number of features were visited multiple times and registered under different numbers. Spatial duplicates and triplicates are mounds that share the same location (often are within 15 m of one another) but have unique ID.

We have two lists of matching TRAP IDs for 2009-2010 mounds and their post-2010 duplicates.
```{r duplicate-lists}
# Early records
upto2010 <- c(6011,8022:8025, 8028,8029,8030, 8035, 8350:8353, 8357,8359, 8434,8669, 9077)

# Later records
post2010 <- c(9357,9594,9595,9593,9596,9592,9591,9358, 8202,9226,9227,9258,9257,9220,9219,9216,9740,9715)

# to see the pairs, they are collated in output data folder
duplicates <- read.csv("../output_data/interim/duplicates_final.txt", sep = " ")

```

In order to highlight the duplicates and allow users to filter the early or later variant (2009-2010 or post-2010), we create a `SpatialDuplicate` column, declare the paired TRAP ID in the `PairedID` column, and create a `Pre2010Variant` and `Post2010Variant` columns to facilitate filtering of the early and later version.  

```{r spatial-duplicate}
# Label the spatial versions/duplicates as such
Y_features$SpatialDuplicate <- ifelse(
  Y_features$TRAP %in% duplicates$upto2010 | 
    Y_features$TRAP %in% duplicates$post2010,
  "Duplicate",
  "Unique"
)

```
Add `PairedID` column and fill in with other version TRAP id where applicable

```{r paired-duplicate-ID}
# Initialize PairedID column with NA
Y_features$PairedID <- NA

# For upto2010 records, use the TRAP id in post2010 column of duplicates
Y_features$PairedID[Y_features$TRAP %in% duplicates$upto2010] <- duplicates$post2010[match(
  Y_features$TRAP[Y_features$TRAP %in% duplicates$upto2010],
  duplicates$upto2010
)]

# For post2010 records, use TRAP id from upto2010 from duplicates
Y_features$PairedID[Y_features$TRAP %in% duplicates$post2010] <- duplicates$upto2010[match(
  Y_features$TRAP[Y_features$TRAP %in% duplicates$post2010],
  duplicates$post2010
)]

Y_features %>% 
  st_drop_geometry() %>% 
  dplyr::select(TRAP, PairedID, SpatialDuplicate) %>% 
  filter(SpatialDuplicate == "Duplicate")

```

Create two columns `Upto2010` and `Post2010` to facilitate filtering by different versions.
```{r filter-duplicates}
`%nin%` <- Negate(`%in%`)
Y_features$Upto2010 <- ifelse(Y_features$TRAP %nin% duplicates$post2010, "Yes", "No")
Y_features$Post2010 <- ifelse(Y_features$TRAP %nin% duplicates$upto2010, "Yes", "No")

# review
Y_features %>% 
  st_drop_geometry() %>% 
  dplyr::select(TRAP, PairedID, SpatialDuplicate, Upto2010, Post2010) %>% 
  filter(SpatialDuplicate == "Duplicate")
```



## Export Yambol features  

```{r export-features}
# Features in Yambol
names(Y_features)
glimpse(Y_features)

# Export
Y_features %>%
  write_rds("../output_data/Y_features.rds")
Y_features %>%
  st_write("../output_data/Y_features.geojson", append = F)

Y_features %>%
  mutate(Easting = st_coordinates(.)[,2],
         Northing = st_coordinates(.)[,1]) %>% 
  mutate(DistBG_m = as.vector(unclass(DistBG_m)), 
         DistTown_m = unclass(DistTown_m),
         DistTownBoundary_m = unclass(DistTownBoundary_m)) %>% 
  st_drop_geometry() %>% 
  write_csv("../output_data/Y_features.csv")
```

## Filter mounds
Now that the attributes look reasonably well, let's filter out and export the most likely mounds inside the Yambol Province.

```{r filter-mounds}
Y_features %>%
   filter(grepl("Mound|Other|Uncertain", Type))%>% 
   group_by(Type) %>% 
   tally()
```


## Export Yambol mounds
Moving to export, .rds, .geojson, and .csv are generated below. For reuse, remember to filter for either the `Upto2010` or `Post2010` variant. The latter is perhaps more authoritative as the later observations are more "current".  But there is good reason to use the former variant if you need to add AKB information or both variants together if you wish to look for change in observations between teams and seasons.

```{r export-mnds}
Y_features %>%
    filter(grepl("[Mm]ound", Type)) %>%
    write_rds("../output_data/Y_mounds.rds")

Y_features %>%
    filter(grepl("[Mm]ound", Type)) %>%
    st_write("../output_data/Y_mounds.geojson", append=FALSE)

Y_features %>%
  filter(grepl("[Mm]ound", Type)) %>%
  mutate(Easting = st_coordinates(.)[,2],
         Northing = st_coordinates(.)[,1]) %>% 
  mutate(DistBG_m = as.vector(unclass(DistBG_m)), 
         DistTown_m = unclass(DistTown_m),
         DistTownBoundary_m = unclass(DistTownBoundary_m)) %>% 
  st_drop_geometry() %>% 
  write_csv("../output_data/Y_mounds.csv")
```
